requirements.txt
anyio==4.2.0
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
Flask==3.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
python-dotenv==1.0.1
python-telegram-bot==20.7
sniffio==1.3.0
Werkzeug==3.0.1
